Scarcity for a single community

Given a stacked data.frame compute species scarcity. Scarcity measures how abundant is a species locally. Scarcity is close to 1 when a species is rare in a community and close to 0 when it is abundant. See scarcity() function or the functional rarity indices vignette included in the package (type vignette("rarity_indices", package = "funrar")) for details about the index.

scarcity_com(com_df, sp_col, abund)

Arguments

com_df
a stacked (= tidy) data.frame from a single community with each row representing a species in a community

sp_col
a character vector, the name of the species column in com_df

abund
a character vector, the name of the column containing relative abundances values

Value

the same data.frame with the additional Si column giving scarcity values for each species

Caution

This function is meant for internal uses mostly, thus it does not include any tests on inputs and may fail unexpectedly. Please use scarcity_stack() to avoid input errors.

See also